+++ azdev/operations/linter/linter.py
@@ -4,17 +4,25 @@
 # license information.
 # -----------------------------------------------------------------------------
 
-import os
-import inspect
-from importlib import import_module
-from pkgutil import iter_modules
+from difflib import context_diff
 from enum import Enum
+from importlib import import_module
+import inspect
+import os
+import re
+from pkgutil import iter_modules
 import yaml
 from knack.log import get_logger
 
+from azdev.operations.regex import (
+    get_all_tested_commands_from_regex,
+    search_argument,
+    search_argument_context,
+    search_command,
+    search_command_group)
+from azdev.utilities import diff_branches_detail
 from azdev.utilities.path import get_cli_repo_path, get_ext_repo_paths
 from .util import share_element, exclude_commands, LinterError
-
 
 PACKAGE_NAME = 'azdev.operations.linter'
 _logger = get_logger(__name__)
@@ -37,8 +45,9 @@ class LinterSeverity(Enum):
         return sorted(LinterSeverity, key=lambda sev: sev.value)
 
 
-class Linter:  # pylint: disable=too-many-public-methods
-    def __init__(self, command_loader=None, help_file_entries=None, loaded_help=None):
+class Linter:  # pylint: disable=too-many-public-methods, too-many-instance-attributes
+    def __init__(self, command_loader=None, help_file_entries=None, loaded_help=None, git_source=None, git_target=None,
+                 git_repo=None, exclusions=None):
         self._all_yaml_help = help_file_entries
         self._loaded_help = loaded_help
         self._command_loader = command_loader
@@ -50,6 +59,10 @@ class Linter:  # pylint: disable=too-many-public-methods
             self._parameters[command_name] = set()
             for name in command.arguments:
                 self._parameters[command_name].add(name)
+        self.git_source = git_source
+        self.git_target = git_target
+        self.git_repo = git_repo
+        self.exclusions = exclusions
 
     @property
     def commands(self):
@@ -180,19 +193,179 @@ class Linter:  # pylint: disable=too-many-public-methods
             return help_entry.short_summary or help_entry.long_summary
         return help_entry
 
+    def get_command_test_coverage(self):
+        diff_index = diff_branches_detail(repo=self.git_repo, target=self.git_target, source=self.git_source)
+        commands, _ = self._detect_new_command(diff_index)
+        all_tested_command = self._detect_tested_command(diff_index)
+        return self._run_command_test_coverage(commands, all_tested_command)
+
+    def get_parameter_test_coverage(self):
+        diff_index = diff_branches_detail(repo=self.git_repo, target=self.git_target, source=self.git_source)
+        _, parameters = self._detect_new_command(diff_index)
+        all_tested_command = self._detect_tested_command(diff_index)
+        return self._run_parameter_test_coverage(parameters, all_tested_command)
+
+    # pylint: disable=too-many-locals, too-many-nested-blocks, too-many-branches, too-many-statements
+    def _detect_new_command(self, diff_index):
+        """
+        exclude_comands: List[str]
+        exclude_parameters: List[tuple[str, str]]
+        commands: List[str]
+        parameters: List[str, List[str]]
+        """
+        YELLOW = '\x1b[33m'
+        parameters = []
+        commands = []
+        lines = []
+        exclude_comands = []
+        exclude_parameters = []
+        for c, v in self.exclusions.items():
+            if 'parameters' in v:
+                for p, r in v['parameters'].items():
+                    if 'missing_parameter_test_coverage' in r['rule_exclusions']:
+                        exclude_parameters.append((c, p))
+            elif 'rule_exclusions' in v:
+                if 'missing_command_test_coverage' in v['rule_exclusions']:
+                    exclude_comands.append(c)
+        _logger.debug('exclude_parameters: %s', exclude_parameters)
+        _logger.debug('exclude_comands: %s', exclude_comands)
+
+        for diff in diff_index:
+            filename = diff.a_path.split('/')[-1]
+            if 'params' in filename or 'commands' in filename:
+                lines = list(
+                    context_diff(diff.a_blob.data_stream.read().decode("utf-8").splitlines(True) if diff.a_blob else [],
+                                 diff.b_blob.data_stream.read().decode("utf-8").splitlines(True) if diff.b_blob else [],
+                                 'Original', 'Current'))
+            for row_num, line in enumerate(lines):
+                if 'params.py' in filename:
+                    params, param_name = search_argument(line)
+                    if params:
+                        offset = -1
+                        while row_num > 0:
+                            row_num -= 1
+                            # Match row num '--- 156,163 ----'
+                            sub_pattern = r'--- (\d{0,}),(?:\d{0,}) ----'
+                            idx = re.findall(sub_pattern, lines[row_num])
+                            offset += 1
+                            if idx:
+                                idx = int(idx[0]) + offset
+                                break
+                        with open(os.path.join(get_cli_repo_path(), diff.a_path), encoding='utf-8') as f:
+                            param_lines = f.readlines()
+                        cmds = search_argument_context(idx, param_lines)
+                        for cmd in cmds:
+                            if cmd not in exclude_comands and \
+                                    not list(filter(lambda x, c=cmd, p=param_name: c in x[0] and p in x[1], exclude_parameters)):  # pylint: disable=line-too-long
+                                parameters.append([cmd, params])
+                            else:
+                                print('%sCommand [%s, %s] not test and exclude in linter_exclusions.yml' % (
+                                    YELLOW, cmd, params))
+
+                if 'commands.py' in filename:
+                    command = search_command(line)
+                    if command:
+                        offset = -1
+                        while row_num > 0:
+                            row_num -= 1
+                            # Match row num '--- 156,163 ----'
+                            sub_pattern = r'--- (\d{0,}),(?:\d{0,}) ----'
+                            idx = re.findall(sub_pattern, lines[row_num])
+                            offset += 1
+                            if idx:
+                                idx = int(idx[0]) + offset
+                                break
+                        with open(os.path.join(get_cli_repo_path(), diff.a_path), encoding='utf-8') as f:
+                            cmd_lines = f.readlines()
+                        cmd = search_command_group(idx, cmd_lines, command)
+                        if cmd:
+                            if cmd in exclude_comands:
+                                print('%sCommand %s not test and exclude in linter_exclusions.yml' % (YELLOW, cmd))
+                            else:
+                                commands.append(cmd)
+        _logger.debug('New add parameters: %s', parameters)
+        _logger.debug('New add commands: %s', commands)
+        return commands, parameters
+
+    def _detect_tested_command(self, diff_index):
+        all_tested_command = []
+        # get tested command by regex
+        for diff in diff_index:
+            filename = diff.a_path.split('/')[-1]
+            if re.findall(r'test_.*.py', filename) and os.path.exists(os.path.join(get_cli_repo_path(), diff.a_path)):
+                with open(os.path.join(self.git_repo, diff.a_path), encoding='utf-8') as f:
+                    lines = f.readlines()
+                ref = get_all_tested_commands_from_regex(lines)
+                all_tested_command += ref
+            # get tested command by recording file
+            if re.findall(r'test_.*.yaml', filename) and os.path.exists(os.path.join(get_cli_repo_path(), diff.a_path)):
+                with open(os.path.join(self.git_repo, diff.a_path)) as f:
+                    records = yaml.load(f, Loader=yaml.Loader) or {}
+                    for record in records['interactions']:
+                        # parse command ['acr agentpool create']
+                        command = record['request']['headers'].get('CommandName', [''])[0]
+                        # parse argument ['-n -r']
+                        argument = record['request']['headers'].get('ParameterSetName', [''])[0]
+                        if command or argument:
+                            all_tested_command.append(command + ' ' + argument)
+        _logger.debug('All tested command: %s', all_tested_command)
+        return all_tested_command
+
+    @staticmethod
+    def _run_command_test_coverage(commands, all_tested_command):
+        exec_state = True
+        violations = []
+        for command in commands:
+            for code in all_tested_command:
+                if command in code:
+                    break
+            else:
+                violations.append(f'Missing command test coverage: `{command}`')
+                exec_state = False
+        if violations:
+            violations.insert(0, 'Failed.')
+            violations.extend([
+                'Please add some scenario tests for the new command',
+                'Or add the command with missing_command_test_coverage rule in linter_exclusions.yml'])
+        return exec_state, violations
+
+    @staticmethod
+    def _run_parameter_test_coverage(parameters, all_tested_command):
+        flag = False
+        exec_state = True
+        violations = []
+        for command, opt_list in parameters:
+            for opt in opt_list:
+                for code in all_tested_command:
+                    if command in code and opt in code:
+                        flag = True
+                        break
+                else:
+                    violations.append(f'Missing parameter test coverage: `{command} {opt}`')
+                    exec_state = False
+                if flag:
+                    break
+        if violations:
+            violations.insert(0, 'Failed.')
+            violations.extend([
+                'Please add some scenario tests for the new parameter',
+                'Or add the parameter with missing_parameter_test_coverage rule in linter_exclusions.yml'])
+        return exec_state, violations
+
 
 # pylint: disable=too-many-instance-attributes
 class LinterManager:
-
-    _RULE_TYPES = {'help_file_entries', 'command_groups', 'commands', 'params'}
+    _RULE_TYPES = {'help_file_entries', 'command_groups', 'commands', 'params', 'command_test_coverage'}
 
     def __init__(self, command_loader=None, help_file_entries=None, loaded_help=None, exclusions=None,
-                 rule_inclusions=None, use_ci_exclusions=None, min_severity=None, update_global_exclusion=None):
+                 rule_inclusions=None, use_ci_exclusions=None, min_severity=None, update_global_exclusion=None,
+                 git_source=None, git_target=None, git_repo=None):
         # default to running only rules of the highest severity
         self.min_severity = min_severity or LinterSeverity.get_ordered_members()[-1]
-        self.linter = Linter(command_loader=command_loader, help_file_entries=help_file_entries,
-                             loaded_help=loaded_help)
         self._exclusions = exclusions or {}
+        self.linter = Linter(command_loader=command_loader, help_file_entries=help_file_entries,
+                             loaded_help=loaded_help, git_source=git_source, git_target=git_target, git_repo=git_repo,
+                             exclusions=self._exclusions)
         self._rules = {rule_type: {} for rule_type in LinterManager._RULE_TYPES}  # initialize empty rules
         self._ci_exclusions = {}
         self._rule_inclusions = rule_inclusions
@@ -234,7 +407,8 @@ class LinterManager:
     def exit_code(self):
         return self._exit_code
 
-    def run(self, run_params=None, run_commands=None, run_command_groups=None, run_help_files_entries=None):
+    def run(self, run_params=None, run_commands=None, run_command_groups=None,
+            run_help_files_entries=None, run_command_test_coverage=None):
         paths = import_module('{}.rules'.format(PACKAGE_NAME)).__path__
 
         if paths:
@@ -256,9 +430,11 @@ class LinterManager:
 
         # run all rule-checks
         if run_help_files_entries and self._rules.get('help_file_entries'):
+            # print('help_file_entries')
             self._run_rules('help_file_entries')
 
         if run_command_groups and self._rules.get('command_groups'):
+            # print('command_groups')
             self._run_rules('command_groups')
 
         if run_commands and self._rules.get('commands'):
@@ -266,6 +442,9 @@ class LinterManager:
 
         if run_params and self._rules.get('params'):
             self._run_rules('params')
+
+        if run_command_test_coverage and self._rules.get('command_test_coverage'):
+            self._run_rules('command_test_coverage')
 
         if not self.exit_code:
             print(os.linesep + 'No violations found for linter rules.')
@@ -297,13 +476,18 @@ class LinterManager:
         YELLOW = '\x1b[33m'
         CYAN = '\x1b[36m'
         RESET = '\x1b[39m'
+        # print('enter _run_rules')
         for rule_name, (rule_func, linter_callable, rule_severity) in self._rules.get(rule_group).items():
+            # print('enter_items')
             severity_str = rule_severity.name
             # use new linter if needed
             with LinterScope(self, linter_callable):
+                # print('enter_with')
                 # if the rule's severity is lower than the linter's severity skip it.
                 if self._linter_severity_is_applicable(rule_severity, rule_name):
+                    # print('enter violations', rule_func)
                     violations = sorted(rule_func()) or []
+                    # print('enter to find')
                     if violations:
                         if rule_severity == LinterSeverity.HIGH:
                             sev_color = RED
@@ -314,13 +498,14 @@ class LinterManager:
 
                         # pylint: disable=duplicate-string-formatting-argument
                         print('- {} FAIL{} - {}{}{} severity: {}'.format(RED, RESET, sev_color,
-                                                                         severity_str, RESET, rule_name,))
+                                                                         severity_str, RESET, rule_name, ))
                         for violation_msg, entity_name, name in violations:
                             print(violation_msg)
                             self._save_violations(entity_name, name)
                         print()
                     else:
                         print('- {} pass{}: {} '.format(GREEN, RESET, rule_name))
+        # print('enter_end')
 
     def _linter_severity_is_applicable(self, rule_severity, rule_name):
         if self.min_severity.value > rule_severity.value:
@@ -336,7 +521,9 @@ class LinterManager:
             self._violiations.setdefault(command_name, {}).setdefault('rule_exclusions', []).append(rule_name)
         else:
             command_name, param_name = entity_name
-            self._violiations.setdefault(command_name, {}).setdefault('parameters', {}).setdefault(param_name, {}).setdefault('rule_exclusions', []).append(rule_name)
+            self._violiations.setdefault(command_name, {}).setdefault('parameters', {}).setdefault(param_name,
+                                                                                                   {}).setdefault(
+                'rule_exclusions', []).append(rule_name)
 
 
 class RuleError(Exception):
@@ -351,6 +538,7 @@ class LinterScope:
     Linter Context manager. used when calling a rule function. Allows substitution of main linter for a linter
     that takes into account any applicable exclusions, if applicable.
     """
+
     def __init__(self, linter_manager, linter_callable):
         self.linter_manager = linter_manager
         self.linter = linter_callable()